% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/allGenerics.R, R/aggDb.R
\docType{class}
\name{aggdb}
\alias{aggdb}
\alias{listSamples}
\alias{getUnit}
\alias{listParams}
\alias{aggdb-class}
\title{Class to manage interactions with an aggdb}
\description{
The \code{\link{aggregate}} method saves genotype aggregates compressed in an aggdb file.
The aggdb class manages connecting/interacting with these files and retrieving
units of interest.
}
\section{Build an aggdb}{

\itemize{
\item \code{aggregate(x, ...)}: Returns an aggregate of genotypes for each individual.
See \code{\link[=aggregate]{aggregate()}} for details.
}
}

\section{Connect to an aggdb}{

\itemize{
\item \code{aggdb(path)}: Connect to an aggdb object.
}
}

\section{Getters}{

In the following code snippets, x is an aggdb object.
\itemize{
\item \code{getUnit(x, unit)}: Retrieve aggregates for specified unit(s).
Use \code{listUnits(x)} to list the units includes in the aggdb
Output will be a matrix.
\item \code{listUnits(x)}: Return a vector of all units included in the aggdb
\item \code{listSamples(x)}: Return a vector of all sample IDs included in the aggdb
}
}

\section{Association testing}{

An aggdb can be directly supplied to the \code{\link[=assocTest]{assocTest()}} method.
}

\section{Merging}{

Aggdbs can be merged using the \code{\link{mergeAggDbs}} method.
}

\examples{
library(rvatData)
gdb <- create_example_gdb()

# generate the aggregates based on a varSetFile
varsetfile <- varSetFile(rvat_example("rvatData_varsetfile.txt.gz"))
varset <- getVarSet(varsetfile, unit = c("NEK1", "SOD1", "ABCA4"), varSetName = "High")
aggfile <- tempfile()
aggregate(x = gdb,
          varSet = varset,
          maxMAF = 0.001,
          output = aggfile,
          verbose = FALSE)

# connect to aggdb, see ?aggdb for more details
aggdb <- aggdb(aggfile)

# list units and samples in aggdb 
head(listUnits(aggdb))
head(listSamples(aggdb))

# retrieve aggregates 
aggregates <- getUnit(aggdb, unit = "SOD1")

# see ?`assocTest-aggdb` for details on running association tests on an aggdb

}
\seealso{
\code{\link{aggregate}}

\code{\link{mergeAggDbs}}

\code{\link{aggdbList}}

\code{\link{assocTest-aggdb}}
}
\keyword{aggdb}
