% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/aggDb-list.R
\docType{class}
\name{aggdbList}
\alias{aggdbList}
\alias{aggdbList-class}
\alias{listUnits,aggdbList-method}
\alias{listSamples,aggdbList-method}
\alias{length,aggdbList-method}
\alias{metadata,aggdbList-method}
\title{Class to facilitate merging aggdbs}
\description{
Class to facilitate merging aggdbs. By providing a vector of aggdb filepaths,
aggdbList will check whether identical samples are included and if duplicated units are included.
The aggdbList can then be used to merge the aggdbs into either a new aggdb,
or merge all included aggregates into a single aggregate score per sample (\code{\link{mergeAggDbs}}).
}
\section{Initialize an aggdbList object}{

\itemize{
\item \code{aggdbList(filelist, checkDups=TRUE)}: Here \code{filelist} is a vector of \code{\link{aggdb}}
filepaths. \code{checkDups} is set to \code{TRUE} by default, in which case an error raised if
unit names are duplicated across aggdbs
}
}

\section{Getters}{

In the following code snippets, x is an aggdbList object.
\itemize{
\item \code{listUnits(x)}: Return a vector of all units included across aggdbs in the aggdbList
\item \code{listSamples(x)}: Return a vector of all samples included across aggdbs in the aggdbList
}
}

\section{Merge or collapse aggdbs}{

\itemize{
\item \code{mergeAggDbs(object, output = NULL, verbose = TRUE)}:
Merge aggregrateFiles, this will generate a new \code{aggdb} including all aggregates across provided aggdbs.
See \code{\link{mergeAggDbs}} for details.
\item \code{collapseAggDbs(object, output = NULL, verbose = TRUE)}: Collapse aggdbs by aggregating values across aggdbs.
This will result in one aggregate score for each sample, representing the aggregate value across aggregate files.
The output will be a two-column matrix including sample IDs and aggregate scores respectively.
See \code{\link{collapseAggDbs}} for details.
}
}

\examples{
library(rvatData)
gdb <- gdb(rvat_example("rvatData.gdb"))

# generate two aggregate files
varsetfile <- varSetFile(rvat_example("rvatData_varsetfile.txt.gz"))
aggdb1 <- tempfile()
aggregate(x = gdb,
          varSet = getVarSet(varsetfile, unit = c("SOD1", "FUS"), varSetName = "High"),
          maxMAF = 0.001,
          output = aggdb1,
          verbose = FALSE)

aggdb2 <- tempfile()
aggregate(x = gdb,
          varSet = getVarSet(varsetfile, unit = c("NEK1"), varSetName = "High"),
          maxMAF = 0.001,
          output = aggdb2,
          verbose = FALSE)

# merge using mergeAggDbs
aggdb <- tempfile()
agglist <- aggdbList(c(aggdb1, aggdb2))
mergeAggDbs(
  agglist,
  output = aggdb
  )

}
\seealso{
\code{\link{mergeAggDbs}}

\code{\link{collapseAggDbs}}

\code{\link{assocTest-aggdb}}

\code{\link{aggregate}}
}
\keyword{aggdbList}
