% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{aggregate,gdb-method}
\alias{aggregate,gdb-method}
\title{Aggregate genotypes into a single (burden) score for each individual}
\usage{
\S4method{aggregate}{gdb}(
  x,
  cohort = "SM",
  varSet = NULL,
  VAR_id = NULL,
  pheno = NULL,
  memlimit = 1000,
  geneticModel = "allelic",
  imputeMethod = "meanImpute",
  MAFweights = "none",
  checkPloidy = NULL,
  keep = NULL,
  output = NULL,
  signif = 3,
  minCallrateVar = 0,
  maxCallrateVar = Inf,
  minCallrateSM = 0,
  maxCallrateSM = Inf,
  minMAF = 0,
  maxMAF = 1,
  minMAC = 0,
  maxMAC = Inf,
  minCarriers = 0,
  maxCarriers = Inf,
  minCarrierFreq = 0,
  maxCarrierFreq = Inf
)
}
\arguments{
\item{x}{a \code{\link{gdb}} object}

\item{cohort}{If a valid cohort name is provided, then the uploaded data for this cohort is used to filter and annotate the genotypes
If not specified, all samples in the gdb will be loaded.}

\item{varSet}{a \code{\link{varSetList}} or \code{\link{varSetFile}} object.}

\item{VAR_id}{A vector of VAR_ids, alternatively the varSet parameter can be specified.
The \code{memlimit} argument controls how many variants to aggregate at a time.}

\item{pheno}{colData field to test as response variable, although not used within this method,
this can be useful to filter samples which have missing data for the response variable.}

\item{memlimit}{Maximum number of variants to load at once (if \code{VAR_id} is specified).}

\item{geneticModel}{Which genetic model to apply? ('allelic', 'recessive' or 'dominant').
Defaults to \code{allelic}.}

\item{imputeMethod}{Which imputation method to apply? ('meanImpute' or 'missingToRef').
Defaults to \code{meanImpute}.}

\item{MAFweights}{MAF weighting method. Currently Madsen-Browning ('mb') is implemented, by default no MAF weighting is applied.}

\item{checkPloidy}{Version of the human genome to use when assigning variant ploidy (diploid,XnonPAR,YnonPAR).
Accepted inputs are GRCh37, hg19, GRCh38, hg38.
If no value is provided then all variants are assigned the default ploidy of "diploid".}

\item{keep}{vector of sample IDs to keep, defaults to \code{NULL}, in which case all samples are kept.}

\item{output}{Output file path for results.
Defaults to \code{NULL}, in which case results are not written.}

\item{minCallrateVar}{Minimum genotype rate for variant retention.}

\item{maxCallrateVar}{Maximum genotype rate for variant retention.}

\item{minCallrateSM}{Minimum genotype rate for sample retention.}

\item{maxCallrateSM}{Maximum genotype rate for sample retention.}

\item{minMAF}{Minimum minor allele frequency for variant retention.}

\item{maxMAF}{Maximum minor allele frequency for variant retention.}

\item{minMAC}{Minimum minor allele count for variant retention.}

\item{maxMAC}{Maximum minor allele count for variant retention.}

\item{minCarriers}{Minimum carrier count for variant retention.}

\item{maxCarriers}{Maximum carrier count for variant retention.}

\item{minCarrierFreq}{Minimum carrier frequency for variant retention.}

\item{maxCarrierFreq}{Maximum carrier frequency for variant retention.}
}
\value{

}
\description{
Returns an aggregate of genotypes for each individual.
Note, the \code{\link{gdb}} implementation is described here, \code{aggregate} can also be run directly on a
\code{\link{genoMatrix}} object as described in the \code{\link{genoMatrix}} documentation.
Specified genetic model, weights, MAF-weighting are taken into account when aggregating.
Aggregates are written to disk in the \code{\link{aggregateFile}} format, which can be used as input
for \code{\link{assocTest-aggregateFile}} to perform gene set burden analyses.
}
