% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{assocTest}
\alias{assocTest}
\title{Perform association tests on binary or quantitative traits}
\value{
The object returned is of class \code{\link{rvbResult}} or \code{\link{singlevarResult}}.
The columns are described below: (note that some columns are specific to either \code{rvbResult} or \code{singlevarResult}
respectively)
\itemize{
\item \code{unit}:  (rvbResult) name of the unit tested (from \code{varSet} or from \code{genoMatrix} metadata)
\item \code{VAR_id}:  (singlevarResult) VAR_id of the respective variant.
\item \code{cohort}:  Name of the cohort
\item \code{varSetName}:  varSetName included in the provided \code{varSet}, if provided, or from the \code{genoMatrix} metadata.
\item \code{name}:  if specified, contains the \code{name} argument specified in assocTest.
\item \code{pheno}:  Phenotype tested, as specified by the \code{pheno} argument.
\item \code{covar}:  Covariates included, as specified by the \code{covar} argument.
\item \code{geneticModel}:  genetic model used, as specified by the \code{geneticModel} argument.
\item \code{MAFweight}: MAF-weighting applied, as specified by the \code{MAFweights} argument.
\item \code{test}:  statistical test used, as specified by the \code{test} argument.
\item \code{nvar}:  (rvbResult) number of variants included in the aggregate test.
\item \code{caseMAC}:  (singlevarResult) Minor allele count among cases
\item \code{ctrlMAC}:  (singlevarResult) Minor allele count among controls.
\item \code{caseCarriers}:  (rvbResult) The number of cases that carry at least one variant among variants in the variant set.
\item \code{ctrlCarriers}:  (rvbResult) The number of controls that carry at least one variant among variants in the variant set.
\item \code{meanCaseScore}:  (rvbResult) Mean burden score among cases.
\item \code{meanCtrlScore}:  (rvbResult) Mean burden score among controls.
\item \code{caseN}:  Number of cases.
\item \code{ctrlN}:  Number of controls.
\item \code{caseCallRate}:  Variant call-rate in cases. For aggregate tests this is the average call-rate across variants.
\item \code{ctrlCallRate}:  Variant call-rate in controls For aggregate tests this is the average call-rate across variants.
\item \code{effectAllele}:  (singlevarResult) Allele to which the effect estimate refers.
\item \code{otherAllele}:  (singlevarResult) Non-effect allele
\item \code{effect}:  Effect estimate for specified statistical test.
Note that SKAT tests, ACAT-v tests, SPA tests and negative binomial tests don't yield effect estimates.
\item \code{effectSE}:  Standard error of effect estimate of specified statistical test.
Note that SKAT tests, ACAT-v tests, SPA tests and negative binomial tests don't yield effect estimates.
\item \code{effectCIlower}:  Lower confidence interval of effect estimate of specified statistical test.
Note that SKAT tests, ACAT-v tests, SPA tests and negative binomial tests don't yield effect estimates.
\item \code{effectCIupper}:  Upper confidence interval of effect estimate of specified statistical test.
Note that SKAT tests, ACAT-v tests, SPA tests and negative binomial tests don't yield effect estimates.
\item \code{OR}:  Odds-ratio for \code{glm} and \code{firth} tests,
the \code{effect} represents the odds-ratio.
\item \code{P}:  P-value for specified statistical test.
}
}
\description{
Main method for performing association tests on binary or quantitative traits.
Performs either aggregate tests or single variant tests, with a wide range of
available statistical methods (see 'Details' for more information).
Various parameters allow for implementation of for example recessive/dominant analyses,
MAF-weighted burden tests and permutation-based association tests.
Variant and sample filters (e.g. maximum MAF, minimum number of carriers) can be specified.
assocTest can be run either directly on a \code{\link{genoMatrix}} object or alternatively
on a \code{\link{gdb}} object, in which case assocTest can loop through multiple variant sets by providing
a \code{\link{varSetFile}} or \code{\link{varSetList}} as input.
Moreover, assocTest can be run on sets of genes ('gene set burden'),
for which gene burden scores have been generated using \code{\link{aggregate}} and stored in an \code{\link{aggregateFile}}.
For running assocTest on a \code{\link{genoMatrix}} object see: \code{\link{assocTest-genoMatrix}} for details\cr
for running assocTest on a \code{\link{gdb}}  object see: \code{\link{assocTest-gdb-method}} for details
for running assocTest on a \code{\link{aggregateFile}}  object see: \code{\link{assocTest-aggregateFile}} for details
}
\details{
\strong{Aggregate tests}
For aggregate tests, the following tests are implemented:

\emph{binary/quantitative traits}:
\itemize{
\item \code{skat}: SKAT test as implemented in the \code{SKAT} R package (Wu \emph{et. al}, 2011).
\item \code{skat_burden}: Burden test as implemented in the \code{SKAT} R package (Wu \emph{et. al}, 2011).
\item \code{skato}: SKAT-O (Optimal Unified Test) as implemented in the \code{SKAT} R package (Wu \emph{et. al}, 2011).
\item \code{acatv}: ACAT-V test as implemented in the \code{ACAT} R package (Liu \emph{et al.}, 2019)
}

\emph{quantitative traits}:
\itemize{
\item \code{lm}: linear model
}

\emph{binary traits}:
\itemize{
\item \code{firth}: firth logistic regression (Firth, 1993). The maximum number of iterations
can be specified using the \code{maxitFirth} parameter
\item \code{glm}: logistic regression
\item \code{nbinom}: Negative binomial test
\item \code{skat_robust}: robust SKAT test (robust in the presence of an unbalanced case/control ratio) as implemented in the \code{SKAT} R package (Zhao \emph{et al.}, 2020).
\item \code{skat_burden_robust}: robust burden test (robust in the presence of an unbalanced case/control ratio) as implemented in the \code{SKAT} R package (Zhao \emph{et al.}, 2020).
\item \code{skato_robust}: robust SKAT-O test (robust in the presence of an unbalanced case/control ratio) (Zhao \emph{et al.}, 2020)
\item \code{acatvSPA}: adjusted ACAT-v test, using a score test using saddlepoint approximation.
}

\strong{Single variant tests}

For single variants, the following tests are implemented:

\emph{quantitative traits}:
\itemize{
\item \code{lm}: linear model
}

\emph{binary traits:}
\itemize{
\item \code{scoreSPA}: score test with saddle point approximation.
\item \code{firth}: firth logistic regression (Firth, 1993). The maximum number of iterations
can be specified using the \code{maxitFirth} parameter.
\item \code{glm}: logistic regression
}
}
\references{
Wu, M. C. et al. Rare-Variant Association Testing for Sequencing Data with the Sequence Kernel Association Test. The American Journal of Human Genetics 89, 82–93 (2011).

Liu, Y. et al. ACAT: A Fast and Powerful p Value Combination Method for Rare-Variant Analysis in Sequencing Studies. The American Journal of Human Genetics 104, 410–421 (2019).

Firth, D. Bias Reduction of Maximum Likelihood Estimates. Biometrika 80, 27–38 (1993).

Zhao, Z. et al. UK Biobank Whole-Exome Sequence Binary Phenome Analysis with Robust Region-Based Rare-Variant Test. The American Journal of Human Genetics 106, 3–12 (2020).
}
