% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{buildResamplingFile}
\alias{buildResamplingFile}
\title{Build a resampling matrix}
\usage{
buildResamplingFile(
  nSamples,
  nResampling = 1000,
  memlimit = 1000,
  methodResampling = "permutation",
  output = NULL
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nResampling}{Number of resamplings. Defaults to 1000.}

\item{memlimit}{Maximum number of resamplings to generate at a time (chunk size).
Relevant when writing to \code{output} to manage memory usage. Defaults to 1000.}

\item{methodResampling}{Resampling method, currently 'permutation' is implemented.}

\item{output}{File path (.gz extension) to write output to.
If not specified, a matrix with resamplings is returned.}
}
\value{
If \code{output} is \code{NULL}, returns a matrix of dimensions \code{nSamples} x \code{nResampling}.
If \code{output} is specified, the function writes to the file and returns \code{NULL}.
}
\description{
Currently the 'permutation' approach is implemented. If \code{output} is specified,
the resampling matrix will be written to disk, and can be connected to with \code{\link{resamplingFile}},
otherwise a matrix is returned. Can be used in combination with \code{\link{assocTest}} to perform
resampling tests.
}
\examples{
library(rvatData)
# build and connect to a resamplingFile
file <- tempfile(fileext = ".gz")
buildResamplingFile(nSamples = 25000, 
                    nResampling = 100,
                    output = file)
resamplingfile <- resamplingFile(file)

# perform resampled association tests 
gdb <- create_example_gdb()
assoc <- assocTest(
  gdb,
  VAR_id = 1:10,
  cohort = "pheno",
  pheno = "pheno",
  covar = paste0("PC", 1:4),
  test = c("skat"),
  resamplingFile = resamplingfile,
  verbose = FALSE
)
}
