% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSet.R
\name{buildVarSet-data.frame}
\alias{buildVarSet-data.frame}
\alias{buildVarSet,data.frame-method}
\title{buildVarSet-data.frame}
\usage{
\S4method{buildVarSet}{data.frame}(
  object,
  unitName,
  fields,
  output = NULL,
  memlimit = NULL,
  mask = TRUE
)
}
\arguments{
\item{object}{A data.frame containing variant annotations}

\item{unitName}{Field to utilize for aggregation unit names.}

\item{fields}{Which fields in the data.frame to use as variant annotations.
These fields can be 1) an indicator (0,1 or FALSE/TRUE) that flags variants with the annotation,
e.g. a column named 'LOF' that indicates whether the variant is predicted to lead to loss-of-function.;
or 2) variant weights, e.g. a column named 'CADD' that contains CADD scores.
Multiple fields can be specified, which will result in multiple rows per aggregation unit in the resulting varSetFile.}

\item{output}{Optional output file name (output will be gz compressed text).}

\item{memlimit}{Defaults to \code{NULL}, currently not implemented.}

\item{mask}{If the annotation field is a 0,1 or FALSE/TRUE indicator, should variants with 0/FALSE be dropped? Defaults to \code{TRUE}.
If \code{FALSE}, all variants will be included, and will be assigned weights 0 and 1.}
}
\description{
Generate optionally weighted variant sets using annotation table(s).
See the tutorials for examples.
}
