% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{buildVarSet}
\alias{buildVarSet}
\title{Generate weighted variant sets for use in association testing.}
\usage{
buildVarSet(object, ...)
}
\arguments{
\item{gdb}{a \code{\link{gdb}} object.}

\item{varSetName}{Name to assign varSet grouping. This identifier column is used to allow for subsequent mergeing of multiple varSet files for coordinated analysis of multiple variant filtering/ weighting strategies)}

\item{unitTable}{Table containing aggregation unit mappings.}

\item{unitName}{Field to utilize for aggregation unit names.}

\item{intersect}{Additional tables to filter through intersection (ie variants absent from intersection tables will not appear in output). Multiple tables should be ',' delimited.}

\item{where}{An SQL compliant where clause to filter output; eg: "CHROM=2 AND POS between 5000 AND 50000 AND AF<0.01 AND (cadd.caddPhred>15 OR snpEff.SIFT='D')".}

\item{weightName}{Field name for desired variant weighting, must be a column within unitTable or other intersection table. Default value of 1 is equivalent to no weighting.}

\item{output}{Output file name (output will be gz compressed text).}
}
\description{
Generate weighted variant sets using annotation table(s) upload to the gdb.
See tutorials for examples.
}
