% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{collapseAggDbs}
\alias{collapseAggDbs}
\title{Collapse multiple aggregate files}
\usage{
collapseAggDbs(object, output = NULL, overWrite = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{an \code{\link{aggdbList}} object.}

\item{output}{Output file name (output will be gz compressed text).
Defaults to \code{NULL}, in which case a data.frame will be returned.}

\item{overWrite}{Should existing output files be overwritten? Defaults to \code{TRUE}.}

\item{verbose}{Should the function be verbose? Defaults to \code{TRUE}.}
}
\description{
Collapse aggregrateFiles by aggregating values across aggdbs. This will result in one aggregate score for
each sample, representing the aggregate value across aggregate dbs.
The output will be a two-column matrix including sample IDs and aggregate scores respectively.
}
\examples{
library(rvatData)
gdb <- create_example_gdb()

# generate two aggregate files
varsetfile <- varSetFile(rvat_example("rvatData_varsetfile.txt.gz"))
aggdb1 <- tempfile()
aggregate(x = gdb,
          varSet = getVarSet(varsetfile, unit = c("SOD1", "FUS"), varSetName = "High"),
          maxMAF = 0.001,
          output = aggdb1,
          verbose = FALSE)

aggdb2 <- tempfile()
aggregate(x = gdb,
          varSet = getVarSet(varsetfile, unit = c("NEK1"), varSetName = "High"),
          maxMAF = 0.001,
          output = aggdb2,
          verbose = FALSE)

# collapse aggdbs
aggdb <- tempfile()
collapseAggDbs(
  aggdbList(c(aggdb1, aggdb2)),
  output = aggdb
)
aggregates <- read.table(aggdb, header = TRUE)
head(aggregates)
}
