% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdb-utils.R
\name{concatGdb}
\alias{concatGdb}
\title{Concatenate gdb databases}
\usage{
concatGdb(
  targets,
  output,
  skipRemap = FALSE,
  skipIndexes = FALSE,
  overWrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{targets}{Character vector of gdb file paths, or a single path to a file listing gdb paths.}

\item{output}{Output gdb file path.}

\item{skipRemap}{Skip resetting of VAR_id to row id after concatenation? Defaults to \code{FALSE}.}

\item{skipIndexes}{Skip generation of standard var and dosage table indexes (VAR_id;CHROM, POS,REF,ALT)?
Defaults to \code{FALSE}.}

\item{overWrite}{Overwrite output file if it already exists?
Defaults to \code{FALSE}.}

\item{verbose}{Should the function be verbose? Defaults to \code{TRUE}.}
}
\description{
Function to concatenate \code{\link{gdb}} databases. Only retains content of base tables (SM, var, dosage).
}
\examples{
library(rvatData)
gdb <- create_example_gdb()

# to illustrate how concatGdb we'll first generate two small gdbs to concatenate
gdb1 <- tempfile()
subsetGdb(
  gdb,
  VAR_id = 1:100,
  output = gdb1
)

gdb2 <- tempfile()
subsetGdb(
  gdb,
  VAR_id = 101:200,
  output = gdb2
)

# write filepaths of gdbs to concatenate to a file
concatgdb <- tempfile()

# concatenate
concatGdb(
  targets = c(gdb1, gdb2),
  output = concatgdb,
)
}
