% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{dropTable}
\alias{dropTable}
\title{Drop table from gdb}
\usage{
dropTable(object, name)
}
\arguments{
\item{object}{\code{\link{gdb}} object.}

\item{name}{Name of table to drop.}
}
\description{
Drop table from \code{\link{gdb}} and clear from annotation / cohort metadata tables.
}
