% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/gdb.R
\name{extractRanges}
\alias{extractRanges}
\alias{extractRanges,gdb-method}
\title{Extract variants from a gdb based on a set of ranges}
\usage{
extractRanges(object, ranges, return = "VAR_id")
}
\arguments{
\item{object}{a \code{\link{gdb}} object}

\item{ranges}{Can be a data.frame, including at least 'CHROM','start', and 'end' columns or
a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object.}

\item{return}{What to return. Defaults to \code{VAR_id}, which returns a vector of VAR_ids that map to specified ranges.}
}
\description{
Extract variant IDs from a \code{\link{gdb}} based on a set of ranges.
}
