% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/nullmodelGSA.R
\name{fitNullModelGSA}
\alias{fitNullModelGSA}
\alias{fitNullModelGSA,rvatResult-method}
\title{fitNullModelGSA}
\usage{
fitNullModelGSA(
  object,
  cormatrix = NULL,
  covar = NULL,
  Zcutoffs = NULL,
  INT = FALSE,
  method = c("GENESIS"),
  ...
)

\S4method{fitNullModelGSA}{rvatResult}(
  object,
  cormatrix = NULL,
  covar = NULL,
  Zcutoffs = NULL,
  INT = FALSE,
  method = c("GENESIS"),
  ...
)
}
\arguments{
\item{object}{\code{\link{rvbResult-class}} object}

\item{cormatrix}{a correlation matrix with row and column names corresponding to the units in the rvbResult. Can be a sparse matrix. See \code{\link{buildCorMatrix}}.}

\item{covar}{a vector of covariates to include when fitting the null model.}

\item{Zcutoffs}{A vector (length=2, minimum and maximum) of cutoffs to apply to the Z-scores.
Z scores below/above these cutoffs will be set equal to the cutoff.}

\item{INT}{Apply inverse normal transformation to Z-scores? Defaults to \code{FALSE}.}

\item{method}{Method to use to fit the mixed linear model, currently 'GENESIS' is implemented.}
}
\description{
Fit a null model for a mixed linear model gene set analysis.
The correlation matrix used to model gene-gene correlations can be generated
using the \code{\link{buildCorMatrix}} method.
}
