% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{forestplot}
\alias{forestplot}
\title{forestplot
Makes a forest plot of the effect size of all occurrences of the chosen gene set}
\usage{
forestplot(object, unit, class = NULL)
}
\arguments{
\item{object}{\code{\link{rvatResult}} object}

\item{unit}{string or vector with names of the units (rvbResult or singlevarResult) or gene sets (gsaResult) of interest}

\item{class}{the type of rvatResult object: rvbResult, singlevarResult, or gsaResult.}
}
\description{
forestplot
Makes a forest plot of the effect size of all occurrences of the chosen gene set
}
