% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/allGenerics.R, R/gdb.R
\docType{class}
\name{gdb}
\alias{gdb}
\alias{listAnno}
\alias{listCohort}
\alias{getRvatVersion}
\alias{getGdbId}
\alias{getGdbPath}
\alias{getGenomeBuild}
\alias{getCreationDate}
\alias{gdb-class}
\alias{show,gdb-method}
\alias{listAnno,gdb-method}
\alias{listCohort,gdb-method}
\alias{getRvatVersion,gdb-method}
\alias{getGdbId,gdb-method}
\alias{getGdbPath,gdb-method}
\alias{getCreationDate,gdb-method}
\alias{getGenomeBuild,gdb-method}
\alias{dropTable,gdb-method}
\title{Connect and interact with a gdb}
\description{
Compressed SQLite (".gdb") representation of genotype data and associated variant annotation and sample info tables
These allow for rapid and memory-efficient of loading sample genotype data en associated metadata within R.
The slots of the gdb class are inherited entirely from the RSQLite SQLiteConnection class.
A host of RVAT methods described here allow for convenient querying and manipulation of a gdb, for complex queries
users can also directly perform SQL queries on the gdb as exemplified in the examples and tutorials.
}
\section{Connect}{

\itemize{
\item \code{gdb(x)}: Connect to a  gdb object  (created using \code{\link[=buildGdb]{buildGdb()}}), where \code{x} is the filepath.
}
}

\section{Build}{

In the following code snippets, x is a gdb object.
\itemize{
\item \code{buildGdb(x)}: build a gdb directly from a vcf, see \code{\link[=buildGdb]{buildGdb()}} for details.
}
}

\section{Getters}{

In the following code snippets, x is a gdb object.
\itemize{
\item \code{getGT(x, ...)}: Retrieve genotype data from the gdb, returns a \code{\link[=genoMatrix]{genoMatrix()}} object.
See \code{\link[=getGT]{getGT()}} for details.
\item \code{getAnno(x, ...)}: Get an annotation table from gdb. See the \code{\link[=getAnno]{getAnno()}} documentation for details.
\item \code{getCohort(x, ...)}: Get a cohort table from gdb. See the \code{\link[=getCohort]{getCohort()}} documentation for details.
\item \code{listCohort(x)}: List sample cohort tables that have been uploaded to the gdb.
\item \code{listAnno(x)}: List variant info tables that have been uploaded to the gdb.
\item \code{getGdbId(x)}, \code{getGdbPath(x)}, \code{getCreationDate(x)}, \code{getGenomeBuild(x)}: various methods to retrieve indicate metadata from gdb
}
}

\section{Upload and delete sample and variant info tables}{

In the following code snippets, x is a gdb object.
\itemize{
\item \code{uploadCohort(x, ...)}: Upload cohort data tables to gdb.
See the \code{\link[=uploadCohort]{uploadCohort()}} documentation for details.
\item \code{uploadAnno(x, ...)}: Upload variant annotation data into gdb.
See the \code{\link[=uploadAnno]{uploadAnno()}} documentation for details
\item \code{mapVariants(x, ...)}: Map variants in the gdb onto features provided in a bed,gff/gtf or ranges file.
See \code{\link[=mapVariants]{mapVariants()}} for details.
\item \code{dropTable(x, name)}: Drop table with the name specified in \code{name} from gdb and clear from annotation / cohort metadata tables.
}
}

\section{Subsetting & Merging}{

\itemize{
\item \code{subSetGdb}: Create a new gdb that is a subset of the input gdb. See \code{\link[=subsetGdb]{subsetGdb()}} for details.
\item \code{concatGdb}: Function to concatenate gdb databases. See \code{\link[=concatGdb]{concatGdb()}} for details.
}
}

\section{Exports}{

\itemize{
\item \code{writeVcf}: Convert a gdb to a vcf-file. See \code{\link[=writeVcf]{writeVcf()}} for details.
}
}

\seealso{
\code{\link{getGT}}

\code{\link{buildGdb}}

\code{\link{getAnno}}

\code{\link{getCohort}}

\code{\link{uploadCohort}}

\code{\link{uploadAnno}}

\code{\link{mapVariants}}

\code{\link{subsetGdb}}

\code{\link{concatGdb}}

\code{\link{writeVcf}}
}
\keyword{gdb}
