% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/geneSetAssoc.R
\docType{class}
\name{geneSet}
\alias{geneSet}
\alias{geneSet-class}
\alias{metadata,geneSet-method}
\alias{length,geneSet-method}
\alias{as.data.frame,geneSet-method}
\alias{listUnits,geneSet-method}
\alias{listWeights,geneSet-method}
\title{Class that represents a geneSet}
\description{
An S4 class to manage an individual geneSet record. Usually multiple geneSets
will be included in a \code{\link{geneSetList}} or \code{\link{geneSetFile}} object, which can be generated using the \code{\link{buildGeneSet}} method.
}
\section{Slots}{

\describe{
\item{\code{geneSetName}}{name of the geneset}

\item{\code{units}}{units (usually genes) in the geneSet}

\item{\code{w}}{optional weights (defaults to 1 for each unit)}

\item{\code{metadata}}{optional metadata for each geneset}
}}

\section{Getters}{

In the following code snippets, x is a geneSet object.
\itemize{
\item \code{listUnits(x)}: Return a vector of all units included in the geneSet
\item \code{listWeights(x)}: Return a vector of all weights in the geneSet
}
}

\seealso{
\code{\link{geneSetFile}}

\code{\link{geneSetList}}

\code{\link{buildGeneSet}}
}
\keyword{geneSet}
