% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/allGenerics.R, R/geneSetAssoc.R
\docType{class}
\name{geneSetFile}
\alias{geneSetFile}
\alias{as.geneSetList}
\alias{geneSetFile-class}
\alias{show,geneSetFile-method}
\alias{names,geneSetFile-method}
\alias{listGeneSets,geneSetFile-method}
\alias{length,geneSetFile-method}
\alias{metadata,geneSetFile-method}
\alias{as.geneSetList,geneSetFile-method}
\title{Class to manage interactions with a geneSetFile}
\usage{
geneSetFile(path, memlimit = 5000)
}
\arguments{
\item{path}{Path to geneSet file.}

\item{memlimit}{Maximum number of records to load at a time.}
}
\description{
An S4 class to manage interactions with geneSetFiles. A geneSetFile can be generated from gmt-files or a list of gene
sets using the \code{\link[=buildGeneSet]{buildGeneSet()}} method. Specific geneSets can be retrieved using
the \code{\link{getGeneSet}} method. A geneSetFile can be used as input in \code{\link[=geneSetAssoc]{geneSetAssoc()}} or \code{\link[=assocTest]{assocTest()}}
to perform competitive/self-contained gene set analyses.

Initialize \code{\link{geneSetFile-class}} object
}
\section{Build a geneSetFile}{

\itemize{
\item \code{buildGeneSet(x, ...)}: Generate a \code{\link{geneSetList}} or geneSetFile that stores (weighted) gene sets
for use geneSetAnalyses. This can be based on 1) GMT-files (https://www.gsea-msigdb.org/gsea/msigdb/) and 2) a list of #' vectors of units per gene set.
See \code{\link[=buildGeneSet]{buildGeneSet()}} for details.
}
}

\section{Connect to a geneSetFile}{

\itemize{
\item \code{geneSetFile(path)}: Connect to a geneSetFile object.
}
}

\section{Getters}{

In the following code snippets, x is a geneSetFile object.
\itemize{
\item \code{getGeneSet(x, geneSet)}: Retrieve geneSets for specified units and/or geneSetNames.
Output will be of class \code{\link{geneSetList}}.
\item \code{listUnits(x)}: Return a vector of all units included in the geneSetFile
\item \code{listGeneSets(x)}: Return a vector of all geneSetNames included in the geneSetFile
}
}

\section{Association testing}{

A geneSetFile can be directly supplied to the \code{\link[=geneSetAssoc]{geneSetAssoc()}} method, using the
\code{geneSet} parameter, in combination with an \code{\link{rvbResult}} object.
To perform gene set burden analyses, \code{\link{assocTest}} can be used.
}

\seealso{
\code{\link{geneSetList}}

\code{\link{buildGeneSet}}

\code{\link{geneSetAssoc}}

\code{\link{assocTest-aggregateFile}}
}
\keyword{geneSetFile}
