% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/allGenerics.R, R/geneSetAssoc.R
\docType{class}
\name{geneSetList}
\alias{geneSetList}
\alias{listMetadata}
\alias{dropUnits}
\alias{remapIDs}
\alias{geneSetList-class}
\alias{names,geneSetList-method}
\alias{length,geneSetList-method}
\alias{lengths,geneSetList-method}
\alias{show,geneSetList-method}
\alias{metadata,geneSetList-method}
\alias{listMetadata,geneSetList-method}
\alias{listGeneSets,geneSetList-method}
\alias{listUnits,geneSetList-method}
\alias{[[,geneSetList,ANY,missing-method}
\alias{[,geneSetList,ANY,ANY,ANY-method}
\alias{sort,geneSetList-method}
\alias{as.data.frame,geneSetList-method}
\alias{as.list,geneSetList-method}
\alias{dropUnits,geneSetList-method}
\alias{remapIDs,geneSetList-method}
\alias{write,geneSetList-method}
\title{Class to manage multiple geneSets}
\description{
An S4 class to manage geneSets. A geneSetList can be generated from gmt-files or a list of gene sets using
the \code{\link[=buildGeneSet]{buildGeneSet()}} method. Specific geneSets can be retrieved using
the \code{\link{getGeneSet}} method. A geneSetList can be used directly as input in \code{\link[=geneSetAssoc]{geneSetAssoc()}} or \code{\link{assocTest}} to perform gene set analyses.
The on-disk equivalent of a \code{geneSetList} is \code{\link[=geneSetFile]{geneSetFile()}}, from which geneSets can be retrieved
in the same way as a geneSetList using the \code{getGeneSet} method.
}
\section{Build a geneSetList}{

\itemize{
\item \code{buildGeneSet(x, ...)}: Generate a geneSetList or \code{\link[=geneSetFile]{geneSetFile()}} that stores (weighted) gene sets
for use geneSetAnalyses. This can be based on 1) GMT-files (https://www.gsea-msigdb.org/gsea/msigdb/) and 2) a list of vectors of units per gene set.
See \code{\link[=buildGeneSet]{buildGeneSet()}} for details.
}
}

\section{Getters}{

In the following code snippets, x is a geneSetList object.
\itemize{
\item \code{geneGeneSet(x, set)}: Retrieve geneSets for specified names.
\item \code{listGeneSets(x)}: Return a vector of all gene sets included in the geneSetList
\item \code{listUnits(x)}: Return a vector of all units included across the gene sets in the geneSetList
}
}

\section{Subsetting}{

A geneSetList can be subsetted in the same way as a normal R list,
however, \code{\link{getGeneSet}} is the most convenient way to select genesets (see above).
Some examples, where x is a geneSetList object:
\itemize{
\item \code{x[[i]]}: Return the i'th geneSet
\item \code{x[i:j]}: Return the i'th till j'th geneSets.
}
}

\section{Gene set analyses}{

A geneSetList can be directly supplied to the \code{\link[=geneSetAssoc]{geneSetAssoc()}} method, using the
\code{geneSet} parameter, in combination with an \code{\link{rvbResult}} object.
To perform gene set burden analyses, \code{\link{assocTest}} can be used.
}

\section{Miscellaneous}{

In the following code snippets, x is a geneSetList object:
\itemize{
\item \code{write(x, file = "data", append = "FALSE")}: Write the geneSet to disk, in
the \code{\link{geneSetFile}} format.
\item \code{remapIDs(x,...)}: Remap IDs used in geneSets, see \code{\link{remapIDs}} for details.
\item \code{dropUnits(x, unit = NULL)}: Remove specified units from geneSets included in the geneSetList.
}
}

\seealso{
\code{\link{geneSetFile}}

\code{\link{geneSet}}

\code{\link{geneSetAssoc}}

\code{\link{assocTest-aggregateFile}}
}
\keyword{geneSetList}
