% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/gdb.R
\name{getAnno}
\alias{getAnno}
\alias{getAnno,gdb-method}
\title{Get an annotation table from a gdb}
\usage{
getAnno(
  object,
  table,
  fields = "*",
  left = c(),
  inner = c(),
  VAR_id = c(),
  ranges = NULL,
  where = c()
)
}
\arguments{
\item{object}{an object of class \code{\link[=gdb-class]{gdb}}}

\item{table}{base table to query}

\item{fields}{columns to retain}

\item{left}{left join operations to perform}

\item{inner}{inner join operations to perform}

\item{VAR_id}{retain only variants with matching ID}

\item{ranges}{Extract variants within specified ranges.
Ranges can be specified as a data.frame, including at least 'CHROM','start', and 'end' columns, or
can be a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object.}

\item{where}{An SQL compliant where clause to filter output; eg: "CHROM=2 AND POS between 5000 AND 50000 AND AF<0.01 AND (cadd.caddPhred>15 OR snpEff.SIFT='D')".}
}
\description{
Get an annotation table from a \code{\link[=gdb-class]{gdb}} object.
}
