% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/gdb_anno_cohort.R
\name{getCohort}
\alias{getCohort}
\alias{getCohort,gdb-method}
\title{Get a cohort table from a gdb}
\usage{
getCohort(object, cohort, fields = "*", keepAll = FALSE)

\S4method{getCohort}{gdb}(object, cohort, fields = "*", keepAll = FALSE)
}
\arguments{
\item{object}{An object of class \code{\link{gdb}}.}

\item{cohort}{Name of cohort to get.}

\item{fields}{Columns to retain. Defaults to '*', in which case all fields are loaded.}

\item{keepAll}{Defaults to \code{FALSE}, for internal use.}
}
\description{
Get a cohort table from a \code{\link{gdb}} object.
}
\examples{
library(rvatData)
gdb <- create_example_gdb()

# retrieve a cohort
cohort <- getCohort(gdb, cohort = "pheno")
head(cohort)

# retrieve a cohort, keep specified fields
cohort <- getCohort(gdb, cohort = "pheno", fields = c("IID", "sex", "pheno"))
head(cohort)
}
