% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/geneSetAssoc.R
\name{getGeneSet}
\alias{getGeneSet}
\alias{getGeneSet,geneSetList-method}
\alias{getGeneSet,geneSetFile-method}
\title{Get geneset(s) from a geneSetList/geneSetFile}
\usage{
getGeneSet(object, geneSet = NULL, unit = NULL)
}
\arguments{
\item{object}{a \code{\link{geneSetList}} or \code{\link{geneSetFile}} object}

\item{geneSet}{a vector of geneSets to subset.}

\item{unit}{a vector of units to subset}
}
\description{
Subset a \code{\link{geneSetList}} or \code{\link{geneSetFile}} object based on one or more geneSet names and/or
based on specific units (i.e. subset the geneSets which contain the specified units.)
}
