% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/varSet.R
\name{getRanges}
\alias{getRanges}
\alias{getRanges,varSetFile-method}
\alias{getRanges,varSetList-method}
\title{getRanges}
\usage{
getRanges(
  object,
  gdb,
  output = NULL,
  table = "var",
  CHROM = "CHROM",
  POS = "POS",
  where = c()
)
}
\arguments{
\item{object}{Input \code{\link{varSetFile}} or \code{\link{varSetList}}.}

\item{gdb}{A \code{\link{gdb}} object.}

\item{output}{Output file path. Output will be gz-compressed. Defaults to \code{NULL},
in which case ranges are returned as a data.frame.}

\item{table}{Name of lookup table for extracting variant positions. Defaults to "var".}

\item{CHROM}{If the name of the column in \code{table} that specifies the chromosome is not named \code{CHROM},
the column name can be specified here.}

\item{POS}{If the name of the column in \code{table} that specifies variant position is not named \code{POS},
the column name can be specified here.}

\item{where}{An SQL compliant where clause to filter output; eg: "CHROM=2 AND POS between 5000 AND 50000 AND AF<0.01 AND (cadd.caddPhred>15 OR snpEff.SIFT='D')". Can be either of length 1, in which case the same
where clause is applied for each varSet, or of length equal to the number of varSets in the provided varSetList or varSetFile,
in which case each where clause is applied to the corresponding varSet.}
}
\description{
Retrieve genomic ranges for variant sets.
}
\details{
This function method retrieves genomic ranges (chromosome, start, and end positions) for variant sets defined in a \code{varSetFile} or \code{varSetList}.
It will map positions based on a variant annotations table in provided \code{gdb} (default = "var").
}
\examples{

library(rvatData)
varsetfile <- varSetFile(rvat_example("rvatData_varsetfile.txt.gz"))
gdb <- create_example_gdb()
ranges <- getRanges(varsetfile, gdb = gdb)

}
