% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getVarSet}
\alias{getVarSet}
\title{Retrieve varSet records from a varSetFile or varSetList}
\usage{
getVarSet(object, unit = NULL, varSetName = NULL)
}
\arguments{
\item{object}{A \code{\link{varSetList}} or \code{\link{varSetFile}}.}

\item{unit}{Vector of unit(s) to retrieve
Use \code{\link[=listUnits]{listUnits()}} to see which units are included in the varSetList/varSetFile.}

\item{varSetName}{Vector of unit(s) to retrieve
Use \code{\link[=listVarSets]{listVarSets()}} to see which varSets are included in the varSetList/varSetFile.}
}
\description{
Retrieve varSet record(s) for specified unit(s) and/or varSetName(s).
\code{object} can be either a \code{\link{varSetList}} or a \code{\link{varSetFile}}.
}
\examples{
library(rvatData)

# connect to varsetfile from rvatData package
# to build a varsetfile, see ?buildVarSet
varsetfile <- varSetFile(rvat_example("rvatData_varsetfile.txt.gz"))

# retrieve specific genes, this will return a varSetList (see ?varSetList)
varsets <- getVarSet(varsetfile, unit = c("SOD1", "FUS"))
head(varsets)

# the varsetfile contains multiple records per gene, (High impact, CADD scores etc.)
unique(listVarSets(varsetfile))
# specific varSets can be selected using the `varSetName` parameter
varsets <- getVarSet(varsetfile, 
                     unit = c("SOD1", "FUS"),
                     varSetName = "CADD"
                     )
head(varsets)

# see?varSetFile and ?varSetList for more details on connecting and handling varsetfiles.
# see e.g., ?assocTest and ?aggregate for downstream methods that can loop through varsetfiles and varsetlists.


}
