% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsaResult.R
\name{gsaResult}
\alias{gsaResult}
\title{Create a \code{\link{gsaResult-class}}}
\usage{
gsaResult(object)
}
\arguments{
\item{object}{GSA results as generated by the \code{\link{geneSetAssoc}} method.
Can be either a data.frame or a filepath pointing to the results.}
}
\description{
Create a \code{\link{gsaResult-class}}
}
