% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{listGeneSets}
\alias{listGeneSets}
\title{Return names of geneSets included in a geneSetFile or geneSetList}
\usage{
listGeneSets(object)
}
\arguments{
\item{object}{a \code{\link{geneSetList}} or \code{\link{geneSetFile}} object}
}
\description{
Returns a vector of all geneSets included in a \code{\link{geneSetFile}} or \code{\link{geneSetList}}
}
