% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{listUnits}
\alias{listUnits}
\title{Return the units included in an varSetFile/varSetList/aggregateFile}
\usage{
listUnits(object)
}
\arguments{
\item{object}{An object of class \code{\link{varSetFile}}, \code{\link{varSetList}} or \code{\link{aggregateFile}}.}
}
\description{
Returns a vector of all units included in a \code{\link{varSetFile}}, \code{\link{varSetList}} or \code{\link{aggregateFile}}
}
