% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{listVarSets}
\alias{listVarSets}
\title{Return names of varsets included in a varSetFile or varSetList}
\usage{
listVarSets(object, ...)
}
\description{
Returns a vector of all varSetNames included in the \code{\link{varSetFile}} or \code{\link{varSetList}}
}
