% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{listWeights}
\alias{listWeights}
\title{Return names of weights included in a varSet or geneSet}
\usage{
listWeights(object, ...)
}
\description{
Returns a vector of weights included in the \code{\link{varSet}} or \code{\link{geneSet}}
}
