% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{manhattanGeneSet}
\alias{manhattanGeneSet}
\title{Generate a manhattan plot with a set of genes highlighted}
\usage{
manhattanGeneSet(
  object,
  geneSet,
  geneSetList,
  gsaObject,
  label = "label",
  threshold = NULL,
  labelThreshold = NULL,
  contigs = c(),
  title = ""
)
}
\arguments{
\item{object}{a \code{\link{rvbResult}} object}

\item{geneSet}{a string with the name of a \code{\link{geneSet}} object}

\item{geneSetList}{a list of \code{\link{geneSet}} objects}

\item{gsaObject}{a \code{\link{gsaResult}} object}
}
\description{
Generate a manhattan plot where the genes of the specified gene set are highlighted and the P-values of each GSA of that gene set is given
}
