% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{mapToCDS}
\alias{mapToCDS}
\title{mapToCDS}
\usage{
mapToCDS(
  object,
  gff,
  exonPadding = 12,
  output = NULL,
  gene_id = NULL,
  transcript_id = NULL,
  biotype = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{Input \code{\link{gdb}}.}

\item{gff}{Can be 1) a path to a valid gff- or gtf-file or
2) a GenomicRanges::GRanges object. Generated, for example, by importing an ensembl gtf file.}

\item{exonPadding}{Remap variants within \code{exonPadding} base pairs from the exon border to the border. Defaults to 12.}

\item{output}{Optional output (.gz). If \code{NULL} remapped coordinates will be returned to session.}

\item{gene_id}{Optional vector of gene ids to keep in the gff/gtf file.}

\item{transcript_id}{Optional vector of transcript ids to keep in the gff/gtf file.}

\item{biotype}{Optional vector of biotypes to keep in the gff/gtf file.}

\item{verbose}{Should the function be verbose? Defaults to \code{TRUE}.}
}
\description{
Maps variants from genomic coordinates to CDS (coding sequence) positions.
}
