% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{mergeAggDbs}
\alias{mergeAggDbs}
\title{Merge multiple aggregate files}
\usage{
mergeAggDbs(object, output, overWrite = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{an \code{\link{aggdbList}} object.}

\item{output}{Output file name (output will be an aggdb).
Defaults to \code{NULL}, in which case a data.frame will be returned.}

\item{overWrite}{Should existing output files be overwritten? Defaults to \code{TRUE}.}

\item{verbose}{Should the function be verbose? Defaults to \code{TRUE}.}
}
\description{
Merge aggdbs, this will generate a new \code{aggdb} including all aggregates across provided aggdbs
}
\examples{
library(rvatData)
gdb <- create_example_gdb()

# generate two aggregate files
varsetfile <- varSetFile(rvat_example("rvatData_varsetfile.txt.gz"))
aggdb1 <- tempfile()
aggregate(x = gdb,
          varSet = getVarSet(varsetfile, unit = c("SOD1", "FUS"), varSetName = "High"),
          maxMAF = 0.001,
          output = aggdb1,
          verbose = FALSE)

aggdb2 <- tempfile()
aggregate(x = gdb,
          varSet = getVarSet(varsetfile, unit = c("NEK1"), varSetName = "High"),
          maxMAF = 0.001,
          output = aggdb2,
          verbose = FALSE)

# merge using mergeAggDbs
aggdb <- tempfile()
agglist <- aggdbList(c(aggdb1, aggdb2))
mergeAggDbs(
  agglist,
  output = aggdb
  )
}
