% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/aggregateFile.R
\name{mergeAggregateFiles}
\alias{mergeAggregateFiles}
\alias{mergeAggregateFiles,aggregateFileList-method}
\title{Merge multiple aggregate files}
\usage{
mergeAggregateFiles(object, collapse = TRUE, output = NULL, verbose = TRUE)
}
\arguments{
\item{object}{an \code{\link{aggregateFileList}} object.}

\item{output}{Output file name (output will be gz compressed text).
Defaults to \code{NULL}, in which case a data.frame will be returned.}

\item{verbose}{Should the function be verbose ? Defaults to \code{TRUE}.}

\item{aggregate}{Aggregate values? Defaults to \code{TRUE}.}
}
\description{
Merge aggregrateFiles. Two types of merging can be performed
\enumerate{
\item \code{aggregate=TRUE}: Aggregate values across aggregateFiles. This will result in one aggregate score for
each sample, representing the aggregate value across aggregate files.
The output will be a two-column matrix including sample IDs and aggregate scores respectively.
\item \code{aggregate=FALSE}: Generate a new \code{aggregateFile} including all aggregates across aggregateFiles.
}
}
