% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/allGenerics.R, R/nullmodelGSA.R
\docType{class}
\name{nullModelGSA-class}
\alias{nullModelGSA-class}
\alias{getNullModel}
\alias{getResults}
\alias{getCovar}
\alias{listUnits,nullModelGSA-method}
\alias{length,nullModelGSA-method}
\alias{show,nullModelGSA-method}
\alias{getNullModel,nullModelGSA-method}
\alias{getResults,nullModelGSA-method}
\alias{getCovar,nullModelGSA-method}
\title{An S4 class to store and handle null models for mixed linear model gene set analysis.}
\usage{
\S4method{listUnits}{nullModelGSA}(object)

\S4method{length}{nullModelGSA}(x)

\S4method{show}{nullModelGSA}(object)
}
\arguments{
\item{object}{\code{\link{nullModelGSA-class}} object}

\item{x}{\code{\link{nullModelGSA-class}} object}
}
\description{
An S4 class to store and handle null models for mixed linear model gene set analysis.

An S4 class to store and handle null models for gene set analysis
}
\section{Slots}{

\describe{
\item{\code{nullmodel}}{the null model}

\item{\code{ID}}{ID field}

\item{\code{results}}{the rvbResults the nullmodel was based on}

\item{\code{units}}{units}

\item{\code{covar}}{covariates included in the null model}

\item{\code{method}}{which method was used, currently 'GENESIS' is implemented.}
}}

\section{Generate a GSA null model}{

\code{fitNullModelGSA}: Generate a GSA null model, see \code{\link[=buildCorMatrix]{buildCorMatrix()}} for details.
}

\section{Getters}{

In the following code snippets, x is a nullModelGSA object
\itemize{
\item \code{getCovar(x)}: Returns fixed effects
\item \code{getResults(x)}: Returns input results
\item \code{getNullModel(x)}: Returns the null model
\item \code{listUnits(x)}: Returns the units included
}
}

\seealso{
\code{\link{geneSetAssoc}}

\code{\link{buildCorMatrix}}

\code{\link{assocTest}}
}
