% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{qqplot}
\alias{qqplot}
\title{Generate a qq-plot}
\usage{
qqplot(
  object,
  title = "",
  label = "label",
  threshold = NULL,
  showThreshold = TRUE,
  labelThreshold = NULL,
  cex = 16,
  lambda1000 = FALSE,
  case = NULL,
  control = NULL
)
}
\arguments{
\item{object}{\code{\link{rvatResult}} object.}

\item{title}{Optional title.}

\item{label}{column that significant results should be labeled with.}

\item{threshold}{Provide alternative fwe threshold (bonferroni is applied by default.)}

\item{labelThreshold}{Provide alternative fwe threshold (bonferroni is applied by default.)}

\item{cex}{Font size, defaults to 16.}

\item{lambda1000}{Show lambda1000? Defaults to \code{FALSE}.
If \code{TRUE}, the maximum number of cases and controls as specified in the \code{\link{rvatResult}} object
will be used to calculate lambda1000, unless the number of cases and controls are specified
using the \code{case} and \code{control} parameters.}

\item{case}{Number of cases in association tests.}

\item{control}{Number of controls in association tests.}
}
\description{
Plot a qqplot based on the P-values stored in an \code{\link{rvatResult}} object.
}
