% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetAssoc.R
\name{readGMT}
\alias{readGMT}
\title{Read a gmt-file}
\usage{
readGMT(path, sep = "\\t")
}
\arguments{
\item{path}{File path}

\item{sep}{Delimiter used}
}
\description{
Read a gmt-file and return an object of class \code{\link{geneSetList}}.
}
