% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{recode}
\alias{recode}
\title{Recode genotypes}
\usage{
recode(object, geneticModel, imputeMethod, weights, MAFweights)
}
\arguments{
\item{object}{An object of class \code{\link{genoMatrix}}}

\item{geneticModel}{Accepted values include "allelic", "dominant", "recessive".}

\item{imputeMethod}{Logic for recoding of missing genotypes. The default, "meanImpute" recodes NA to the mean genotype dosage. "missingToRef" recodes NA to 0.}

\item{weights}{A vector of weights, with length equal to the number of variants.}

\item{MAFweights}{Apply MAF-dependent weighting, current options include "none" and "mb" (Madsen-Browning).}
}
\description{
Returns a recoded genoMatrix object, genetic model, imputation, and weights can be recoded.
}
