% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/resampling.R
\docType{class}
\name{resamplingFile}
\alias{resamplingFile}
\alias{resamplingFile-class}
\alias{show,resamplingFile-method}
\title{Generate and manage permutations}
\description{
This class is designed to generate and manage permutations. For some purposes
it is useful to store permutations, for example for empirically establishing the correlation
among genes using permutations. In this case identical permutations should be used across genes,
which can be achieved by storing the permutations in a resamplingFile.
}
\section{Construct a  resamplingFile}{

\code{buildResamplingFile}: Construct a resamplingfile, see \code{\link{buildResamplingFile}} for details.
}

\section{Connect}{

\itemize{
\item \code{resamplingFile(x)}: Connect to a resamplingFile object (created using \code{\link[=buildResamplingFile]{buildResamplingFile()}}), where \code{x} is the filepath.
}
}

\section{Association testing}{

A resamplingFile can be directly provided to \code{\link[=assocTest]{assocTest()}}, to run resampled association tests.
}

\examples{
library(rvatData)
# build and connect to a resamplingFile
file <- tempfile(fileext = ".gz")
buildResamplingFile(nSamples = 25000, 
                    nResampling = 100,
                    output = file)
resamplingfile <- resamplingFile(file)

# perform resampled association tests 
gdb <- gdb(rvat_example("rvatData.gdb"))
assoc <- assocTest(
  gdb,
  VAR_id = 1:10,
  cohort = "pheno",
  pheno = "pheno",
  covar = paste0("PC", 1:4),
  test = c("skat"),
  resamplingFile = resamplingfile,
  verbose = FALSE
)
}
