% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvatViewer.R
\name{rvatViewer}
\alias{rvatViewer}
\title{rvatViewer}
\usage{
rvatViewer(results, CHROM = NULL, POS = NULL)
}
\arguments{
\item{results}{Can be 1) an object of class \code{\link{rvbResult}} or \code{\link{singlevarResult}},
2) a filepath to an SQL-database,
3) a \code{data.frame}, or 4) a list of objects of class \code{\link{rvbResult}}, \code{\link{singlevarResult}}, or \code{data.frame}}

\item{CHROM}{if the name of the column that specifies the chromosome is not named \code{CHROM},
the column name can be specified here.}

\item{POS}{if the name of the column that specifies the chromosomal position is not named \code{POS},
the column name can be specified here.}
}
\description{
Interactively explore results of variant/burden testing
}
