% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{uploadAnno}
\alias{uploadAnno}
\title{Upload variant annotation into gdb.}
\usage{
uploadAnno(
  object,
  name,
  value,
  sep = "\\t",
  skipRemap = FALSE,
  skipIndexes = FALSE,
  ignoreAlleles = FALSE,
  keepUnmapped = FALSE,
  mapRef = "var"
)
}
\arguments{
\item{object}{A \code{\link{gdb}} object.}

\item{name}{Name to assign to annotation table.}

\item{value}{Input variant annotation table. Can be a data frame object or a path to a text file.}

\item{sep}{Field delimiter (only used if value is a text file). Defaults to \verb{\\\\t}.}

\item{skipRemap}{Flag indicating whether to skip mapping of records to VAR_id using CHROM,POS,REF,ALT. Defaults to \code{FALSE}.}

\item{skipIndexes}{Flag indicating whether to skip indexing of imported table. Defaults to \code{FALSE}.}

\item{ignoreAlleles}{Flag indicating whether to consider REF and ALT allele during mapping of records to VAR_id or just CHROM,POS. Defaults to \code{FALSE}.}

\item{keepUnmapped}{Flag indicating whether to keep records which cannot be mapped to the gdb. Defaults to \code{FALSE}.}

\item{mapRef}{Name of lookup table for VAR_id assignment. Defaults to "var".}
}
\description{
Function to upload variant annotation data into gdb. Assignment of VAR_id is performed automatically through mapping to the var table, but can be skipped using the skipRemap option. Indexing of the imported table based on VAR_id is also automated but can also be skipped, this could be desirable if it is intended to concatenate many separately generated gdb as otherwise intermediate indexes are generated unecessarily.
}
