% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/allGenerics.R
\docType{class}
\name{varSet}
\alias{varSet}
\alias{listVars}
\title{Class that represents a set of variants and weights}
\description{
An S4 class to manage an individual varSet record. Usually multiple varSets
will be included in a \code{\link{varSetList}} or \code{\link{varSetFile}} object, which can be generated using the \code{\link{buildVarSet}} method.
}
\section{Slots}{

\describe{
\item{\code{unit}}{Unit name (such as the gene or transcript ID)}

\item{\code{varSetName}}{Name of the variant set (such as 'LOF', 'moderate', 'CADD')}

\item{\code{VAR_id}}{VAR_ids included in the varSet}

\item{\code{w}}{weights included in the varSet}

\item{\code{metadata}}{metadata}
}}

\section{Getters}{

In the following code snippets, x is a varSet object.
\itemize{
\item \code{listVars(x)}: Return a vector of all VAR_ids included in the varSet
\item \code{listWeights(x)}: Return a vector of all weights in the varSet
}
}

\section{Retrieve genotypes}{

A varSet object can be supplied to the \code{varSet} parameter in the \code{\link[=getGT]{getGT()}} method to load the
variants included in the varSet.
}

\seealso{
\code{\link{varSetFile}}

\code{\link{varSetList}}

\code{\link{getGT}}

\code{\link{buildVarSet}}
}
