% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfUtils.R
\name{vcfInfo2Table}
\alias{vcfInfo2Table}
\title{Convert vcf info field to table format}
\usage{
vcfInfo2Table(vcf, output, splitMultiallelic = TRUE)
}
\arguments{
\item{vcf}{Input vcf file path.}

\item{output}{Output file path}

\item{splitMultiallelic}{Returns one row per alternative allele instead of one row per variant.
Note that this option currently simply duplicates the information in the INFO field for each allele;
it does not parse Number='A' or Number='R' INFO fields to extract allele-specific values.
Defaults to \code{TRUE}.}
}
\description{
Convert vcf info field to table format. Requires valid vcf where INFO fields are specified in the header.
}
\examples{
library(rvatData)
vcf <- rvat_example("rvatData.vcf.gz")
output <- tempfile()
vcfInfo2Table(
  vcf = vcf,
  output = output
)
}
