% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/aggregateFile.R
\docType{class}
\name{aggregateFileList}
\alias{aggregateFileList}
\alias{aggregateFileList-class}
\alias{show,aggregateFileList-method}
\alias{listUnits,aggregateFileList-method}
\alias{listSamples,aggregateFileList-method}
\alias{length,aggregateFileList-method}
\alias{metadata,aggregateFileList-method}
\title{Class to facilitate merging aggregateFiles}
\description{
Class to facilitate merging aggregateFiles. By providing a vector of aggregateFile filepaths,
aggregateFileList will check whether identical samples are included and if duplicated units are included.
The aggregateFileList can then be used to merge the aggregateFiles into either a new aggregateFile,
or merge all included aggregates into a single aggregate score per sample (\code{\link{mergeAggregateFiles}}).
}
\section{Initialize an aggregateFileList object}{

\itemize{
\item \code{aggregateFileList(filelist, checkDups=TRUE)}: Here \code{filelist} is a vector of \code{\link{aggregateFile}}
filepaths. \code{checkDups} is set to \code{TRUE} by default, in which case an error raised if
unit names are duplicated across aggregateFiles.
}
}

\section{Getters}{

In the following code snippets, x is an aggregateFileList object.
\itemize{
\item \code{listUnits(x)}: Return a vector of all units included across aggregateFiles in the aggregateFileList.
\item \code{listSamples(x)}: Return a vector of all samples included across aggregateFiles in the aggregateFileList.
}
}

\section{Merge aggregateFiles}{

\itemize{
\item \code{mergeAggregateFiles(object,aggregate = TRUE,output = NULL,verbose = TRUE)}:
Merge the aggregatefiles included in the aggregateFileList.
Either aggregate all aggregates into one single aggregate per sample, or merge
all aggregatefiles into a new aggregateFile. See \code{\link{mergeAggregateFiles}} for details.
}
}

\seealso{
\code{\link{mergeAggregateFiles}}

\code{\link{assocTest-aggregateFile}}

\code{\link{aggregate}}
}
\keyword{aggregateFileList}
