% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocTest_aggregateFile.R
\name{assocTest-aggregateFile}
\alias{assocTest-aggregateFile}
\alias{assocTest,aggregateFile-method}
\title{assoctest-aggregateFile}
\usage{
\S4method{assocTest}{aggregateFile}(
  object,
  pheno,
  test,
  geneSet,
  gdb,
  cohort = "SM",
  name = "none",
  continuous = FALSE,
  covar = NULL,
  substractCovar = NULL,
  dropUnits = NULL,
  maxitFirth = 1000,
  keep = NULL,
  output = NULL,
  verbose = TRUE,
  strict = TRUE
)
}
\arguments{
\item{object}{a \code{\link{aggregateFile}} object (generated using the \code{\link{aggregate}} method).}

\item{pheno}{field to test as response variable, the response variable
can either be binary (0/1) or continuous. If the response variable is continuous set
\code{continuous} to \code{TRUE}.}

\item{test}{Vector of statistical tests to run,
options include firth,glm,lm, and nbinom.}

\item{geneSet}{a \code{\link{geneSetList}} or \code{\link{geneSetFile}} object}

\item{gdb}{a \code{\link{gdb}} object, this will be used to load the cohort data.}

\item{cohort}{Cohort name (present in the gdb)}

\item{name}{Optional name of the analysis.}

\item{continuous}{Is the response variable continuous? (TRUE/FALSE). Defaults to \code{FALSE}.}

\item{covar}{Character vector of covariates, or list with multiple sets of covariates.}

\item{substractCovar}{Covariate from which aggregate should be substracted.
Useful when adjusting for total variant counts, by specifying the total variant count variable
here, the aggregate score of the gene set tested will be substracted from the total count variable.}

\item{dropUnits}{Optional, vector of units to exclude.}

\item{maxitFirth}{Maximum number of iterations to use for estimating firth confidence intervals.}

\item{keep}{vector of sample IDs to keep, defaults to \code{NULL}, in which case all samples are kept.
Defaults to \code{FALSE}.}

\item{output}{Output file path for results.
Defaults to \code{NULL}, in which case results are not written but returned as a \code{data.frame()}.}

\item{verbose}{Should the function be verbose? (TRUE/FALSE), defaults to \code{TRUE}.}

\item{strict}{Should strict checks be performed? Defaults to \code{TRUE}. Strict tests currently includes
checking whether supplied varSetFile/varSetList was generated from the same gdb as specified in \code{object}.}
}
\description{
Run \code{\link{assocTest}} on a \code{\link{aggregateFile}} object.
See the main \code{\link{assocTest}} page for details.
}
