% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/nullmodelGSA.R
\name{buildCorMatrix}
\alias{buildCorMatrix}
\alias{buildCorMatrix,rvatResult,aggregateFile-method}
\title{buildCorMatrix}
\usage{
buildCorMatrix(
  object,
  aggregateFile,
  memlimit = 5000,
  minR2 = 1e-04,
  makePD = TRUE,
  absolute = TRUE,
  maxDist = 2500000,
  verbose = TRUE
)

\S4method{buildCorMatrix}{rvatResult,aggregateFile}(
  object,
  aggregateFile,
  memlimit = 5000,
  minR2 = 1e-04,
  makePD = TRUE,
  absolute = TRUE,
  maxDist = 2500000,
  verbose = TRUE
)
}
\arguments{
\item{object}{\code{\link{rvatResult}} object}

\item{aggregateFile}{\code{\link{aggregateFile}} object}

\item{memlimit}{maximum number of units to load from the aggregateFile at a time.}

\item{minR2}{R2 values < minR2 will be set to zero (leading to increased sparsity)}

\item{makePD}{Make the correlation matrix positive definite? (TRUE/FALSE)}

\item{absolute}{Should cormatrix be absolute? Defaults to \code{TRUE}.}

\item{maxDist}{A distance larger than \code{maxDist} defines a new block. Defaults to 2.5Mb (5Mb window)}

\item{verbose}{Should the function be verbose? Defaults to \code{TRUE}.}
}
\description{
Build a block-wise burden correlation matrix, in order to correct for gene-gene correlations in \code{\link{geneSetAssoc}}.
Burden scores should be stored in an \code{\link{aggregateFile}} object (see \code{\link{aggregate}}).
The size of the blocks are controlled using the \code{maxDist} parameter, all gene-gene correlations beyond the block are set to zero.
This function is based on previous work (\url{https://github.com/opain/TWAS-GSEA}).
}
\references{
\url{https://github.com/opain/TWAS-GSEA}
}
