% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdbUtils.R
\name{buildGdb}
\alias{buildGdb}
\title{Create a gdb file.}
\usage{
buildGdb(
  output,
  vcf = c(),
  skipIndexes = FALSE,
  skipVarRanges = FALSE,
  overWrite = FALSE
)
}
\arguments{
\item{output}{Path for output \code{\link{gdb}} file}

\item{vcf}{Input vcf file used to structure and populate gdb. Warning this function makes the following of assumptions: 1) strict adherence to vcf format (GT subfield first element in genotype firelds), 2) multiallelic records have been split, 3) desired genotype QC has already been applied (DP,GQ filters), 4) GT values conform to the set {0/0,0/1,1/0,1/1,./.,0|0,0|1,1|0,1|1,.|.}. Multiallelic parsing and genotype QC can be performed using vcftools and/or accompanying parser scripts included on the rvat github.}

\item{skipIndexes}{Flag to skip generation of indexes for var and dosage table (VAR_id;CHROM, POS,REF,ALT). Typically only required if you plan to use gdbConcat to concatenate a series of separately generated gdb files before use}

\item{skipVarRanges}{Flag to skip generation of ranged var table. Typically only required if you plan to use gdbConcat to concatenate a series of separately generated gdb files before use}

\item{overWrite}{overwrite if \code{output} already exists? Defaults to \code{FALSE}, in which case an error is raised.}
}
\description{
Creates a new \code{\link{gdb}} file and returns a connection object of type gdb-class.
The gdb can be structured and populated using a provided vcf file.
If no input variant file is provided then only an empty gdb is created.
}
