% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneSetAssoc.R
\name{buildGeneSet}
\alias{buildGeneSet}
\title{Build a geneSetList or geneSetFile}
\usage{
buildGeneSet(
  data = NULL,
  gmtpath = NULL,
  output = NULL,
  sep = "\\t",
  verbose = TRUE
)
}
\arguments{
\item{data}{Can be 1) data.frame where the first column includes the names of geneSets,
the second column contains the genes included in the geneSet (comma-delimited).
The third and fourth column are optional: the third column contains weights (comma-delimited),
and the fourth column contains metadata.
or 2) a list, where the names of the list represent the geneSet names and each element in the list contains
a vector with the genes included in the respective geneset.}

\item{gmtpath}{Path to a gmt-file}

\item{output}{Optional output file path (output will be gz compressed text).
Defaults to \code{NULL}, in which case a \code{\link{geneSetList}} is returned.}

\item{sep}{Separator used in input file.}

\item{verbose}{Should the function be verbose? (TRUE/FALSE), defaults to \code{TRUE}.}
}
\description{
Build a \code{\link{geneSetList}} or \code{\link{geneSetFile}} for use in gene set analyses (\code{\link{geneSetAssoc}} or \code{\link{assocTest-aggregateFile}})
Currently these can be build directly from GMT-files, data.frames and lists.
}
\examples{

# build a genesetlist from a list (see ?geneSetList)
genesetlist <- buildGeneSet(
  list("geneset1" = c("SOD1", "NEK1"),
       "geneset2" = c("ABCA4", "SOD1", "NEK1"),
       "geneset3" = c("FUS", "NEK1")
       ))

# specify the output parameter to write to disk in the geneSetFile format (see ?geneSetFile)
file <- tempfile()
buildGeneSet(
  list("geneset1" = c("SOD1", "NEK1"),
       "geneset2" = c("ABCA4", "SOD1", "NEK1"),
       "geneset3" = c("FUS", "NEK1")
  ),
  output = file
  )
genesetfile <- geneSetFile(file)

# the `gmtpath` parameter can be used to build a geneset from a mSigDb GMT-file
# see the tutorials on the RVAT website for examples

}
