% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{buildResamplingFile}
\alias{buildResamplingFile}
\title{Build a resampling matrix}
\usage{
buildResamplingFile(
  nSamples,
  nResampling,
  memlimit = 1000,
  methodResampling = "permutation",
  output = NULL
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nResampling}{Number of resamplings}

\item{memlimit}{Chunk sizes (when writing to output)}

\item{methodResampling}{Resampling method, currently 'permutation' is implemented.}

\item{output}{File path (.gz extension) to write output to. If not specified, a matrix with resamplings is returned.}
}
\description{
Currently the 'permutation' approach is implemented. If \code{output} is specified,
the resampling matrix will be written to disk, and can be connected to with \code{\link{resamplingFile}},
otherwise a matrix is returned. Can be used in combination with \code{\link{assocTest}} to perform
resampling tests.
}
