% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{buildResamplingFile}
\alias{buildResamplingFile}
\title{Build a resampling matrix}
\usage{
buildResamplingFile(
  nSamples,
  nResampling = 1000,
  memlimit = 1000,
  methodResampling = "permutation",
  output = NULL
)
}
\arguments{
\item{nSamples}{Number of samples}

\item{nResampling}{Number of resamplings. Defaults to 1000.}

\item{memlimit}{Chunk sizes (when writing to output)}

\item{methodResampling}{Resampling method, currently 'permutation' is implemented.}

\item{output}{File path (.gz extension) to write output to. If not specified, a matrix with resamplings is returned.}
}
\description{
Currently the 'permutation' approach is implemented. If \code{output} is specified,
the resampling matrix will be written to disk, and can be connected to with \code{\link{resamplingFile}},
otherwise a matrix is returned. Can be used in combination with \code{\link{assocTest}} to perform
resampling tests.
}
\examples{
library(rvatData)
# build and connect to a resamplingFile
file <- tempfile(fileext = ".gz")
buildResamplingFile(nSamples = 25000, 
                    nResampling = 100,
                    output = file)
resamplingfile <- resamplingFile(file)

# perform resampled association tests 
gdb <- gdb(rvat_example("rvatData.gdb"))
assoc <- assocTest(
  gdb,
  VAR_id = 1:10,
  cohort = "pheno",
  pheno = "pheno",
  covar = paste0("PC", 1:4),
  test = c("skat"),
  resamplingFile = resamplingfile,
  verbose = FALSE
)
}
