% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{buildVarSet}
\alias{buildVarSet}
\title{Generate optionally weighted variant sets using annotation table(s).
See the tutorials for examples.
For building varSets directly from the \code{\link{gdb}}: see \code{\link{buildVarSet-gdb}} for details\cr
For building varSets interactively from a data.frame  seee \code{\link{buildVarSet-data.frame}} for details\cr}
\description{
Generate optionally weighted variant sets using annotation table(s).
See the tutorials for examples.
For building varSets directly from the \code{\link{gdb}}: see \code{\link{buildVarSet-gdb}} for details\cr
For building varSets interactively from a data.frame  seee \code{\link{buildVarSet-data.frame}} for details\cr
}
