% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{dropTable}
\alias{dropTable}
\title{Drop table from gdb}
\usage{
dropTable(object, name, verbose = TRUE)
}
\arguments{
\item{object}{\code{\link{gdb}} object.}

\item{name}{Name of table to drop.}

\item{verbose}{Should the method be verbose? Defaults to \code{TRUE}.}
}
\description{
Drop table from \code{\link{gdb}} and clear from annotation / cohort metadata tables.
}
\examples{
library(rvatData)
gdb <- create_example_gdb()
dropTable(gdb, name = "varInfo")

}
