% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{geneSetAssoc}
\alias{geneSetAssoc}
\title{geneSetAssoc}
\usage{
geneSetAssoc(
  object,
  geneSet = NULL,
  scoreMatrix = NULL,
  cormatrix = NULL,
  condition = NULL,
  covar = NULL,
  test = c("lm", "glm", "mlm", "fisher", "ttest", "ztest", "ACAT"),
  threshold = NULL,
  Zcutoffs = NULL,
  INT = FALSE,
  scoreCutoffs = NULL,
  minSetSize = 1,
  maxSetSize = Inf,
  oneSided = TRUE,
  memlimit = 1000,
  REML = TRUE,
  ID = "unit",
  output = NULL
)
}
\arguments{
\item{object}{an \code{\link{rvbResult}} object.}

\item{geneSet}{a \code{\link{geneSetList}} or \code{\link{geneSetFile}} object.}

\item{scoreMatrix}{a matrix (rows = genes, columns = features)}

\item{cormatrix}{a correlation matrix with row and column names corresponding to the units in the rvbResult.}

\item{condition}{Perform conditional analyses. Input can be a 1) a \code{\link{geneSetList}} or \code{\link{geneSetFile}}, in which case
the genesets specified in the \code{geneSet} parameter will all be conditioned on the gene sets provided here.
2) a vector of gene set names present in \code{geneSet}, all genesets specified in the geneSetList/geneSetFile
will be conditioned on the genesets specified here.}

\item{covar}{a vector of covariates}

\item{test}{Vector of tests to perform.
Currently implemented tests are the competitive tests lm,mlm,fisher and self-contained tests ttest,ztest and ACAT.}

\item{threshold}{A vector of thresholds for cutoff-based tests (fisher's exact test / glm).}

\item{Zcutoffs}{A vector (length=2, minimum and maximum) of cutoffs to apply to the Z-scores.
Z scores below/above these cutoffs will be set equal to the cutoff.}

\item{INT}{Apply inverse normal transformation to Z-scores? Defaults to \code{FALSE}.}

\item{minSetSize}{Exclude genesets with size < minSetSize}

\item{maxSetSize}{Exclude genesets with size > maxSetSize}

\item{oneSided}{Calculate a one-sided P-value? Defaults to \code{TRUE}.}

\item{memlimit}{Maximum number of genesets to process in one go.}

\item{REML}{Use REML for mixed linear models? Defaults to \code{TRUE}.}

\item{ID}{ID column in the rvbResult that corresponds with the IDs used in the geneSetList.
Defaults to 'unit'.}

\item{output}{Optional: save results to specified path}
}
\description{
Perform self-contained or competitive gene set analyses.
}
