% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/gdb.R
\name{getCohort}
\alias{getCohort}
\alias{getCohort,gdb-method}
\title{Get an cohort table from a gdb}
\usage{
getCohort(object, cohort, fields = "*", keepAll = FALSE)

\S4method{getCohort}{gdb}(object, cohort, fields = "*", keepAll = FALSE)
}
\arguments{
\item{object}{an object of class \code{\link[=gdb-class]{gdb}}}

\item{cohort}{name of cohort to get.}

\item{fields}{columns to retain}

\item{keepAll}{defaults to \code{FALSE}, for internal use.}
}
\description{
Get a cohort table from a \code{\link[=gdb-class]{gdb}} object.
}
\examples{
library(rvatData)
gdb <- gdb(rvat_example("rvatData.gdb"))

# retrieve a cohort
cohort <- getCohort(gdb, cohort = "pheno")
head(cohort)

# retrieve a cohort, keep specified fields
cohort <- getCohort(gdb, cohort = "pheno", fields = c("IID", "sex", "pheno"))
head(cohort)

}
