% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/gdb.R
\name{getCohort}
\alias{getCohort}
\alias{getCohort,gdb-method}
\title{Get an cohort table from a gdb}
\usage{
getCohort(object, cohort, fields = "*")

\S4method{getCohort}{gdb}(object, cohort, fields = "*")
}
\arguments{
\item{object}{an object of class \code{\link[=gdb-class]{gdb}}}

\item{cohort}{name of cohort to get.}

\item{fields}{columns to retain}
}
\description{
Get a cohort table from a \code{\link[=gdb-class]{gdb}} object.
}
