% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/gdb.R
\name{getGT}
\alias{getGT}
\alias{getGT,gdb-method}
\title{Load genotypes from a gdb.}
\usage{
getGT(
  object,
  varSet = NULL,
  VAR_id = NULL,
  ranges = NULL,
  cohort = NULL,
  checkPloidy = NULL,
  varSetName = "unnamed",
  unit = "unnamed",
  verbose = TRUE
)

\S4method{getGT}{gdb}(
  object,
  varSet = NULL,
  VAR_id = NULL,
  ranges = NULL,
  cohort = NULL,
  checkPloidy = NULL,
  varSetName = "unnamed",
  unit = "unnamed",
  verbose = TRUE
)
}
\arguments{
\item{object}{an object of class \code{\link[=gdb-class]{gdb}}}

\item{varSet}{\code{\link{varSet}} object. If specified, the \code{VAR_id}, \code{varSetName} and \code{unit} parameters will be ignored.}

\item{VAR_id}{Charaocter vector containing target VAR_id.}

\item{ranges}{Extract variants within specified ranges.
Ranges can be specified as a data.frame, including at least 'CHROM','start', and 'end' columns, or
can be a \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges}} object.}

\item{cohort}{Optional use of cohort data previously uploaded to the gdb (see uploadCohort).
If a valid cohort name is provided, then the uploaded data for this cohort is used to filter and annotate the returned genoMatrix object. If a dataframe is provided, then this is assumed to conform to the SM table constraints required for genoMatrix objects (see genoMatrix).}

\item{checkPloidy}{Version of the human genome to use when assigning variant ploidy (diploid,XnonPAR,YnonPAR).
Accepted inputs are GRCh37, hg19, GRCh38, hg38.
If no value is provided then all variants are assigned the default ploidy of "diploid"}

\item{varSetName}{Optional name for the set of variants, for example: 'missense or 'LOF' (ignored if \code{varSet} is specified.)}

\item{unit}{Optional 'unit' name, for example: 'SOD1' or 'ENSG00000142168' (ignored if \code{varSet} is specified.)}

\item{verbose}{Should the method be verbose? Defaults to \code{TRUE}.}
}
\value{
A \code{\link{genoMatrix}} object.
}
\description{
Method to retrieve a \code{\link{genoMatrix}} for variants specified by \code{varSet} or a vector of VAR_ids,
and samples as specified by the \code{cohort} parameter.
The \code{checkPloidy} parameter can be set to \code{GRCh37} (or \code{hg19}) or \code{GRCh38} (or \code{hg38}) to
assign variant ploidy (diploid,XnonPAR,YnonPAR).
}
