% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{getVarSet}
\alias{getVarSet}
\title{Retrieve varSet records from a varSetFile or varSetList}
\usage{
getVarSet(object, unit = NULL, varSetName = NULL)
}
\arguments{
\item{object}{A \code{\link{varSetList}} or \code{\link{varSetFile}}.}

\item{unit}{Vector of unit(s) to retrieve
Use \code{\link[=listUnits]{listUnits()}} to see which units are included in the varSetList/varSetFile.}

\item{varSetName}{Vector of unit(s) to retrieve
Use \code{\link[=listVarSets]{listVarSets()}} to see which varSets are included in the varSetList/varSetFile.}
}
\description{
Retrieve varSet record(s) for specified unit(s) and/or varSetName(s).
\code{object} can be either a \code{\link{varSetList}} or a \code{\link{varSetFile}}.
}
