% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{listGeneSets}
\alias{listGeneSets}
\title{Return the geneSets included in an geneSetFile/geneSetList}
\usage{
listGeneSets(object)
}
\description{
Returns a vector of all units included in a \code{\link{geneSetFile}} or \code{\link{geneSetList}}
}
