% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/aggregateFile.R
\name{mergeAggregateFiles}
\alias{mergeAggregateFiles}
\alias{mergeAggregateFiles,aggregateFileList-method}
\title{Merge multiple aggregate files}
\usage{
mergeAggregateFiles(object, output = NULL, verbose = TRUE)
}
\arguments{
\item{object}{an \code{\link{aggregateFileList}} object.}

\item{output}{Output file name (output will be gz compressed text).
Defaults to \code{NULL}, in which case a data.frame will be returned.}

\item{verbose}{Should the function be verbose? Defaults to \code{TRUE}.}
}
\description{
Merge aggregrateFiles, this will generate a new \code{aggregateFile} including all aggregates across provided aggregateFiles.
}
