% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{summariseGeno}
\alias{summariseGeno}
\title{Return variant summaries}
\usage{
summariseGeno(object, ...)
}
\arguments{
\item{object}{a \code{\link{gdb}} object}

\item{cohort}{If a valid cohort name is provided, then the uploaded data for this cohort is used to filter and annotate the genotypes
If not specified, all samples in the gdb will be loaded.}

\item{varSet}{a \code{\link{varSetList}} or \code{\link{varSetFile}} object.}

\item{VAR_id}{A list of VAR_ids, alternatively the varSet parameter can be specified.
If single variant tests are ran, the \code{memlimit} argument controls how many variants to analyze at a time.}

\item{pheno}{colData field to test as response variable, although not used within this method,
this can be useful to filter samples which have missing data for the response variable.}

\item{memlimit}{Maximum number of variants to load at once (if \code{VAR_id} is specified).}

\item{geneticModel}{Which genetic model to apply? ('allelic', 'recessive' or 'dominant').
Defaults to \code{allelic}.}

\item{checkPloidy}{Version of the human genome to use when assigning variant ploidy (diploid,XnonPAR,YnonPAR).
Accepted inputs are GRCh37, hg19, GRCh38, hg38.
If no value is provided then all variants are assigned the default ploidy of "diploid".}

\item{keep}{vector of sample IDs to keep, defaults to \code{NULL}, in which case all samples are kept.}

\item{output}{Output file path for results.
Defaults to \code{NULL}, in which case results are not written.}

\item{splitBy}{Split variant summaries by labels indicated in the field
specified by \code{splitBy}.}

\item{minCallrateVar}{Minimum genotype rate for variant retention.}

\item{maxCallrateVar}{Maximum genotype rate for variant retention.}

\item{minCallrateSM}{Minimum genotype rate for sample retention.}

\item{maxCallrateSM}{Maximum genotype rate for sample retention.}

\item{minMAF}{Minimum minor allele frequency for variant retention.}

\item{maxMAF}{Maximum minor allele frequency for variant retention.}

\item{minMAC}{Minimum minor allele count for variant retention.}

\item{maxMAC}{Maximum minor allele count for variant retention.}

\item{minCarriers}{Minimum carrier count for variant retention.}

\item{maxCarriers}{Maximum carrier count for variant retention.}

\item{minCarrierFreq}{Minimum carrier frequency for variant retention.}

\item{maxCarrierFreq}{Maximum carrier frequency for variant retention.}
}
\value{
Returns a \code{data.frame} with the following columns:
\itemize{
\item \code{VAR_id}:  VAR_id of the respective variant.
\item \code{AF}:  Allele frequency
\item \code{callRate}:  callRate
\item \code{geno0}:  Number of samples with genotype='0'.
When \code{geneticModel}='allelic' or 'dominant' this is the number of individuals that are
homozygous for the reference allele.
\item \code{geno1}:  Number of samples with genotype='1'.
When \code{geneticModel}='allelic' this is the number of individuals that are
heterozygous for the reference allele.
When \code{geneticModel} = 'dominant' this represents the number of individuals who carry at least one
alternate allele.
When \code{geneticModel} = 'recessive' this represents the number of individuals who are homozygous
for the alternate allele.
\item \code{geno2}: When \code{geneticModel} = 'allelic', the number of individuals who are homozygous
for the alternate allele.
}
}
\description{
Returns a per variant summary of genotype counts, frequencies, call-rates and hwe testing.
Note, the \code{\link{gdb}} implementation is described here, \code{summariseGeno} can also be run directly on a
\code{\link{genoMatrix}} object as described in the  \code{\link{genoMatrix}} documentation.
}
