% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R, R/varSet.R
\docType{class}
\name{varSetFile}
\alias{varSetFile}
\alias{varSetFile-class}
\title{Class to manage interactions with a varSetFile}
\description{
An S4 class to manage interactions with varSetFiles. varSets can be generated from annotations using
the \code{\link[=buildVarSet]{buildVarSet()}} method. Specific units and/or annotations can be loaded using
the \code{getVarSet} method. A varSetFile can be used as input in \code{\link[=assocTest]{assocTest()}} to perform
burden/single variant association tests on the varSets included in the varSetFile.
}
\section{Build a varSetFile}{

\itemize{
\item \code{buildVarSet(x, ...)}: Generate a varSetList or \code{\link[=varSetFile]{varSetFile()}} that stores weighted variant sets
for use in association testing. This can be based on 1) annotations uploaded to the gdb or 2) a data.frame including annotations.
See \code{\link[=buildVarSet]{buildVarSet()}} for details.
}
}

\section{Connect to a varSetFile}{

\itemize{
\item \code{varSetFile(path, memlimit = 5000)}: Connect to a varSetFile object.
}
}

\section{Getters}{

In the following code snippets, x is a varSetFile object.
\itemize{
\item \code{getVarSet(x, unit, varSetName)}: Retrieve varSets for specified units and/or varSetNames.
Output will be of class \code{\link{varSetList}}.
\item \code{listUnits(x)}: Return a vector of all units included in the varSetFile
\item \code{listVarSets(x)}: Return a vector of all varSetNames included in the varSetFile
}
}

\section{Association testing}{

A varSetFile can be directly supplied to the \code{\link[=assocTest]{assocTest()}} gdb method, using the
\code{varSet} parameter. Association tests will then be performed for each varSet included
in the varSetFile.
}

\seealso{
\code{\link{varSetList}}

\code{\link{buildVarSet}}

\code{\link{assocTest}}
}
\keyword{varSetFile}
