% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfUtils.R
\name{vcfInfo2Table}
\alias{vcfInfo2Table}
\title{Convert vcf info field to table format}
\usage{
vcfInfo2Table(vcf, output, splitMultiallelic = TRUE)
}
\arguments{
\item{vcf}{vcf file path.}

\item{output}{output path}

\item{splitMultiallelic}{Returns one row per alternative allele instead of one row per variant. Default=TRUE.}
}
\description{
Convert vcf info field to table format. Requires valid vcf where INFO fields are specified in header.
}
