% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{writeVcf}
\alias{writeVcf}
\title{writeVcf}
\usage{
writeVcf(object, output, VAR_id, IID, includeGeno = TRUE)
}
\arguments{
\item{object}{Input \code{\link{gdb}}.}

\item{output}{Output file path.}

\item{VAR_id}{Variants to include in output.}

\item{IID}{Samples to include in output.}

\item{includeGeno}{Can be reset to false if individual genotypes are not required in output.
Defaults to \code{TRUE}.}
}
\description{
Takes input and outputs to file in vcf format.
}
