% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/gdb.R
\name{writeVcf}
\alias{writeVcf}
\alias{writeVcf,gdb-method}
\title{writeVcf}
\usage{
writeVcf(
  object,
  output,
  VAR_id = NULL,
  IID = NULL,
  includeGeno = TRUE,
  includeVarId = FALSE,
  verbose = TRUE
)

\S4method{writeVcf}{gdb}(
  object,
  output,
  VAR_id = NULL,
  IID = NULL,
  includeGeno = TRUE,
  includeVarId = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Input \code{\link{gdb}}.}

\item{output}{Output file path.}

\item{VAR_id}{Variants to include in output.}

\item{IID}{Samples to include in output.}

\item{includeGeno}{Can be reset to false if individual genotypes are not required in output.
Defaults to \code{TRUE}.}

\item{includeVarId}{Include VAR_ids in the 'ID' field? Defaults to \code{FALSE},
in which case the 'ID' field from the 'var' table is included.}

\item{verbose}{Should the method be verbose? Defaults to \code{TRUE}.}
}
\description{
Takes input and outputs to file in vcf format.
}
